feat: Add jenkins proxy support for v2.X servers
